Switch from CxList String to CxList Char
The fact is that the imaginary boundary between two consecutive Rights
in a CxList String really just don't mean anything, and we just
haphazardly concat them when we need to. So give up on those.

There are a couple places where HOFs are piling up horrifically, but
those can wait until the morning.

<3 Haskell; also <3 my former self for writing unit tests :P